refactor: Move scoped annotations to data.annotation
