+++ src/winter/data/annotation/Controller.java
@@ -1,4 +1,4 @@
-package winter.data.annotation.methodlevel;
+package winter.data.annotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -6,6 +6,6 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.ElementType;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface Rest {
+@Target(ElementType.TYPE)
+public @interface Controller {
 }
